+++ react-components/src/components/CostsAndPricing/Direct/index.jsx
@@ -12,7 +12,7 @@ export const Direct = memo(({
   const perUnit = `${currency} per unit`
   return (
     <>
-      <h2 className='h-xs p-t-l p-b-0'>Direct costs</h2>
+      <h2 className='h-m p-v-s'>Direct costs</h2>
       <p>These are the costs that go directly into building your product into something sellable. It is important you
         know your direct costs so you don't end up selling your product for less than you paid to make it.</p>
       <p>Record all of your direct costs in the table to calculate your total.</p>