+++ CHANGELOG.md
@@ -138,6 +138,7 @@
 - GP2-315 - route to market backend integration
 
 ### Implemented enhancements
+- GP2-1047 - page structure and components for costs and pricing page  
 - GP2-1033 - Country chooser mobile
 - GP2-1051 - Compare markets feature flag
 - GP2-984 - Compare markets selector and redux global state managememt